Update b2b-source.md
I believe there is a minor issue in the description of `sourceKey`. The current formula shows as `[sourceID]@$[sourceInstanceID].[sourceType]`. The `$` symbol looks incorrect here. As far as I know and in the projects I am working on, the separator is the `@` symbol. The proposed change removes this extra symbol.